help: add plugin commands listing TODO: define usage in plugin scripts and show them in help
